# src/main/resources/openapi/api-cash-online.yml
openapi: 3.0.0
info:
  title: API Business Layer
  version: "0.1"
paths:

  /users/{id}:
    get:
      summary: get users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
            format: long

      operationId: getUsers


      responses:
        '200':
          description: Cards state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGetResponse'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '404':
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'


    delete:
      summary: delete users
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
            format: long

      responses:
        '200':
          description: Cards state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostUserResponse'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '404':
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'




  /users:

    post:

      summary: post users
      operationId: saveUsers

      requestBody:
        description: "User request post"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUserRequest'

      responses:
        '200':
          description: Cards state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostUserResponse'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '404':
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'




components:
  schemas:
    UserGetResponse:
      type: object
      description: ""
      properties:
        email:
          type: string
          pattern: '^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$'
        first_name:
          type: string
        last_name:
          type: string
        loans:
          $ref: '#/components/schemas/loanList'

    loanList:
      type: array
      items:
        $ref: '#/components/schemas/Loan'
      minItems: 1

    Loan:
      type: object
      properties:
        id:
          type: number
          format: long
        total:
          type: number
          format: float
        userId:
          type: number
          format: long

    GeneralBusinessError:
      type: object
      description: ''
      properties:
        code:
          type: string
        message:
          type: string

    PostUserRequest:
      type: object
      description: ""
      properties:
        email:
          type: string
          pattern: '^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$'
        first_name:
          type: string
        last_name:
          type: string

    PostUserResponse:
      type: object
      description: ''
      properties:
        code:
          type: string
        message:
          type: string